requirements.txt
opencv-python
numpy
pytest
pillow-heif
Pillow
ultralytics
flask
